Prevent bundle reconciles triggered by bundle status updates
A bundle must only be reconciled upon changes impacting its spec, or a
related cluster or bundle deployment. Changes to the bundle's status
must not lead to new reconciles, which may cause endless loops.
